module1/requirements.txt
# ML Dependencies
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.0
torch==1.9.0
tensorflow==2.5.0

# VULNERABILITY: Known vulnerable packages (Safety will detect)
flask==0.12.2
requests==2.19.1
pyyaml==5.1
urllib3==1.24.1
jinja2==2.10
werkzeug==0.15.3
cryptography==2.1.4
pillow==6.2.0
django==2.0.0
paramiko==2.4.1

# Crypto (for intentional weak crypto examples)
pycryptodome==3.9.8

# Database
pymongo==3.11.0
redis==3.5.0
sqlalchemy==1.3.0

# AWS
boto3==1.17.0
